+++ lib/src/player/music_metadata.dart
@@ -24,6 +24,30 @@ class MusicMetadata {
   })  : assert(mediaId != null),
         assert(duration != null);
 
+  factory MusicMetadata.fromMap(Map map) {
+    return MusicMetadata(
+      extras: map["extras"],
+      mediaId: map["mediaId"],
+      title: map["title"],
+      subtitle: map["subtitle"],
+      duration: map["duration"],
+      iconUri: map["iconUri"],
+      mediaUri: map['mediaUri'],
+    );
+  }
+
+  Map toMap() {
+    return {
+      "extras": extras,
+      "mediaId": mediaId,
+      "title": title,
+      "subtitle": subtitle,
+      "duration": duration,
+      "iconUri": iconUri,
+      "mediaUri": mediaUri,
+    };
+  }
+
   @override
   String toString() {
     return 'MusicMetadata{mediaId: $mediaId, title: $title, subtitle: $subtitle}';
