+++ lib/src/player/play_queue.dart
@@ -20,5 +20,23 @@ class PlayQueue {
         assert(queueTitle != null),
         assert(queue != null);
 
+  factory PlayQueue.fromMap(Map map) {
+    return PlayQueue(
+      queueId: map['queueId'],
+      queueTitle: map['queueTitle'],
+      extras: map['extras'],
+      queue: (map['queue'] as List).cast<Map>().map((e) => MusicMetadata.fromMap(e)).toList(),
+    );
+  }
+
+  Map toMap() {
+    return {
+      "queueId": queueId,
+      "queueTitle": queueTitle,
+      "extras": extras,
+      "queue": queue.map((e) => e.toMap()).toList(),
+    };
+  }
+
   const PlayQueue.empty() : this(queueId: "empty", queueTitle: "empty", queue: const [], extras: const {});
 }
